docs(faq): remove reference to relaxed queue limit
It seems the relaxed queue limit may have been changed to 3
for consistency with the fast queue. Or these are both sharing
the same concurrency limit between them.

This removes the reference to the relaxed jobs being different.
Seems clear enough now that the 3 limit is across both relaxed
and fast queues since no mention is made to either.